+++ commands.go
@@ -21,7 +21,7 @@ import (
 )
 
 func makeClassPathCmd(conf *t4cConfig, toolsClassPath string) exec.Cmd {
-	cmdArgs := append([]string{}, "-classpath", toolsClassPath,
+	cmdArgs := append([]string{}, os.Getenv("CLJ_JVM_OPTS"), "-classpath", toolsClassPath,
 		"clojure.main",
 		"-m", "clojure.tools.deps.alpha.script.make-classpath2",
 		"--config-user", conf.configUser,
@@ -42,7 +42,7 @@ func makeClassPathCmd(conf *t4cConfig, toolsClassPath string) exec.Cmd {
 }
 
 func generatePomCmd(conf *t4cConfig, toolsClassPath string) exec.Cmd {
-	cmdArgs := append([]string{}, "-classpath", toolsClassPath,
+	cmdArgs := append([]string{}, os.Getenv("CLJ_JVM_OPTS"), "-classpath", toolsClassPath,
 		"clojure.main",
 		"-m", "clojure.tools.deps.alpha.script.generate-manifest2",
 		"--config-user", conf.configUser,
@@ -60,7 +60,8 @@ func generatePomCmd(conf *t4cConfig, toolsClassPath string) exec.Cmd {
 func clojureExecuteCmd(jvmCacheOpts []string, jvmOpts []string, basisFile string,
 	execJarPath string, cp string, args []string) exec.Cmd {
 
-	cmdArgs := append([]string{}, jvmCacheOpts...)
+	cmdArgs := append([]string{}, os.Getenv("JAVA_OPTS"))
+	cmdArgs = append(cmdArgs, jvmCacheOpts...)
 	cmdArgs = append(cmdArgs, jvmOpts...)
 	cmdArgs = append(cmdArgs, "-Dclojure.basis="+basisFile,
 		"-classpath", getExecCpFile(cp, execJarPath))
@@ -77,7 +78,8 @@ func clojureExecuteCmd(jvmCacheOpts []string, jvmOpts []string, basisFile string
 func clojureCmd(jvmCacheOpts []string, jvmOpts []string, libsFile string, basisFile string,
 	cp string, mainCacheOpts []string, clojureArgs []string, rlwrap bool) exec.Cmd {
 
-	cmdArgs := append([]string{}, jvmCacheOpts...)
+	cmdArgs := append([]string{}, os.Getenv("JAVA_OPTS"))
+	cmdArgs = append(cmdArgs, jvmCacheOpts...)
 	cmdArgs = append(cmdArgs, jvmOpts...)
 	cmdArgs = append(cmdArgs, "-Dclojure.libfile="+libsFile, "-Dclojure.basisfile="+basisFile, "-classpath", cp, "clojure.main")
 	cmdArgs = append(cmdArgs, mainCacheOpts...)