+++ commands_test.go
@@ -71,6 +71,43 @@ func TestMakeClassPathCmd(t *testing.T) {
 			}
 		}
 	}
+
+	// test when CLJ_JVM_OPTS environment variable is set
+	os.Setenv("CLJ_JVM_OPTS", "CLJ_JVM_OPTS_VALUE")
+
+	cmd = makeClassPathCmd(&conf, toolsCpDir)
+
+	os.Unsetenv("CLJ_JVM_OPTS")
+
+	// test args
+	{
+		expected := []string{
+			javaPath, "CLJ_JVM_OPTS_VALUE", "-classpath", toolsCpDir,
+			"clojure.main",
+			"-m", "clojure.tools.deps.alpha.script.make-classpath2",
+			"--config-user", conf.configUser,
+			"--config-project", conf.configProject,
+			"--libs-file", conf.libsFile,
+			"--basis-file", conf.basisFile,
+			"--cp-file", conf.cpFile,
+			"--jvm-file", conf.jvmFile,
+			"--main-file", conf.mainFile,
+			"--manifest-file", conf.manifestFile,
+		}
+		expected = append(expected, conf.toolsArgs...)
+		expected = removeEmpty(expected)
+
+		if len(cmd.Args) != len(expected) {
+			t.Errorf("makeClassPathCmd failed, args expected %v, got %v", len(expected), len(cmd.Args))
+			t.FailNow()
+		}
+
+		for i, v := range cmd.Args {
+			if v != expected[i] {
+				t.Errorf("makeClassPathCmd failed, arg expected %v, got %v", expected[i], v)
+			}
+		}
+	}
 }
 
 func TestGeneratePomCmd(t *testing.T) {
@@ -98,6 +135,43 @@ func TestGeneratePomCmd(t *testing.T) {
 	{
 		expected := []string{
 			javaPath, "-classpath", toolsCpDir,
+			"clojure.main",
+			"-m", "clojure.tools.deps.alpha.script.generate-manifest2",
+			"--config-user", conf.configUser,
+			"--config-project", conf.configProject,
+			"--gen=pom",
+		}
+		expected = append(expected, conf.toolsArgs...)
+		expected = removeEmpty(expected)
+
+		if len(cmd.Args) != len(expected) {
+			t.Errorf("generatePomCmd failed, args expected %v, got %v", len(expected), len(cmd.Args))
+
+			for i1, v1 := range cmd.Args {
+				t.Errorf("expected %v = %v", i1, v1)
+			}
+
+			t.FailNow()
+		}
+
+		for i, v := range cmd.Args {
+			if v != expected[i] {
+				t.Errorf("generatePomCmd failed, arg expected %v, got %v", expected[i], v)
+			}
+		}
+	}
+
+	// test when CLJ_JVM_OPTS environment variable is set
+	os.Setenv("CLJ_JVM_OPTS", "CLJ_JVM_OPTS_VALUE")
+
+	cmd = generatePomCmd(&conf, toolsCpDir)
+
+	os.Unsetenv("CLJ_JVM_OPTS")
+
+	// test args
+	{
+		expected := []string{
+			javaPath, "CLJ_JVM_OPTS_VALUE", "-classpath", toolsCpDir,
 			"clojure.main",
 			"-m", "clojure.tools.deps.alpha.script.generate-manifest2",
 			"--config-user", conf.configUser,
@@ -253,6 +327,41 @@ func TestClojureExecuteCmd(t *testing.T) {
 			}
 		}
 	}
+
+	// test clojure -X execute, with aliases, and args, when JAVA_OPTS environment variable is set
+	{
+		args := []string{"-X:foo", "arg1", "arg2"}
+
+		os.Setenv("JAVA_OPTS", "JAVA_OPTS_VALUE")
+
+		cmd := clojureExecuteCmd(jvmCacheOpts, jvmOpts, conf.basisFile,
+			execJarPath, cp, args)
+
+		os.Unsetenv("JAVA_OPTS")
+
+		expected := []string{javaPath}
+
+		expected = append(expected, "JAVA_OPTS_VALUE")
+		expected = append(expected, jvmCacheOpts...)
+		expected = append(expected, jvmOpts...)
+		expected = append(expected, "-Dclojure.basis="+conf.basisFile,
+			"-classpath", cp+string(os.PathListSeparator)+execJarPath)
+		expected = append(expected, "clojure.main", "-m", "clojure.run.exec")
+		expected = append(expected, args...)
+
+		expected = removeEmpty(expected)
+
+		if len(cmd.Args) != len(expected) {
+			t.Errorf("clojureExecuteCmd (-X:foo arg1 arg2) failed, args expected %v, got %v", len(expected), len(cmd.Args))
+			t.FailNow()
+		}
+
+		for i, v := range cmd.Args {
+			if v != expected[i] {
+				t.Errorf("clojureExecuteCmd (-X:foo arg1 arg2) failed, arg expected %v, got %v", expected[i], v)
+			}
+		}
+	}
 }
 
 func TestClojureCmd(t *testing.T) {
@@ -277,12 +386,12 @@ func TestClojureCmd(t *testing.T) {
 		"test",
 	}
 
-	cmd := clojureCmd(jvmCacheOpts, jvmOpts, conf.libsFile, conf.basisFile,
-		cp, mainCacheOpts, clojureArgs, false)
-
 	expected := []string{javaPath}
 
 	// test clojure args
+	cmd := clojureCmd(jvmCacheOpts, jvmOpts, conf.libsFile, conf.basisFile,
+		cp, mainCacheOpts, clojureArgs, false)
+
 	{
 		expected = append(expected, jvmCacheOpts...)
 		expected = append(expected, jvmOpts...)
@@ -303,23 +412,82 @@ func TestClojureCmd(t *testing.T) {
 		}
 	}
 
+	// test clj (rlwrap'ed) args
 	cmd = clojureCmd(jvmCacheOpts, jvmOpts, conf.libsFile, conf.basisFile,
 		cp, mainCacheOpts, clojureArgs, true)
 
-	// test clj (rlwrap'ed) args
 	{
+		expectedWrapped := []string{javaPath}
+
 		if runtime.GOOS != "windows" {
-			expected = append(rlwrapArgs(), expected...)
+			expectedWrapped = append(rlwrapArgs(), expected...)
 		}
 
+		if len(cmd.Args) != len(expectedWrapped) {
+			t.Errorf("rlwrap'ed clojureCmd failed, args expected %v, got %v", len(expectedWrapped), len(cmd.Args))
+			t.FailNow()
+		}
+
+		for i, v := range cmd.Args {
+			if v != expectedWrapped[i] {
+				t.Errorf("rlwrap'ed clojureCmd failed, arg expected %v, got %v", expectedWrapped[i], v)
+			}
+		}
+	}
+
+	// test clojure args, when JAVA_OPTS environment variable is set
+	os.Setenv("JAVA_OPTS", "JAVA_OPTS_VALUE")
+
+	cmd = clojureCmd(jvmCacheOpts, jvmOpts, conf.libsFile, conf.basisFile,
+		cp, mainCacheOpts, clojureArgs, false)
+
+	os.Unsetenv("JAVA_OPTS")
+
+	{
+		expected = []string{javaPath}
+		expected = append(expected, "JAVA_OPTS_VALUE")
+		expected = append(expected, jvmCacheOpts...)
+		expected = append(expected, jvmOpts...)
+		expected = append(expected, "-Dclojure.libfile="+conf.libsFile, "-Dclojure.basisfile="+conf.basisFile, "-classpath", cp, "clojure.main")
+		expected = append(expected, mainCacheOpts...)
+		expected = append(expected, clojureArgs...)
+		expected = removeEmpty(expected)
+
 		if len(cmd.Args) != len(expected) {
-			t.Errorf("rlwrap'ed clojureCmd failed, args expected %v, got %v", len(expected), len(cmd.Args))
+			t.Errorf("clojureCmd failed, args expected %v, got %v", len(expected), len(cmd.Args))
 			t.FailNow()
 		}
 
 		for i, v := range cmd.Args {
 			if v != expected[i] {
-				t.Errorf("rlwrap'ed clojureCmd failed, arg expected %v, got %v", expected[i], v)
+				t.Errorf("clojureCmd failed, arg expected %v, got %v", expected[i], v)
+			}
+		}
+	}
+
+	// test clj (rlwrap'ed) args, when JAVA_OPTS environment variable is set
+	os.Setenv("JAVA_OPTS", "JAVA_OPTS_VALUE")
+
+	cmd = clojureCmd(jvmCacheOpts, jvmOpts, conf.libsFile, conf.basisFile,
+		cp, mainCacheOpts, clojureArgs, true)
+
+	os.Unsetenv("JAVA_OPTS")
+
+	{
+		expectedWrapped := []string{javaPath}
+
+		if runtime.GOOS != "windows" {
+			expectedWrapped = append(rlwrapArgs(), expected...)
+		}
+
+		if len(cmd.Args) != len(expectedWrapped) {
+			t.Errorf("rlwrap'ed clojureCmd failed, args expected %v, got %v", len(expectedWrapped), len(cmd.Args))
+			t.FailNow()
+		}
+
+		for i, v := range cmd.Args {
+			if v != expectedWrapped[i] {
+				t.Errorf("rlwrap'ed clojureCmd failed, arg expected %v, got %v", expectedWrapped[i], v)
 			}
 		}
 	}